receipt

+++ backend/bill/bill/Models/item.cs
@@ -11,7 +11,7 @@ public partial class item
 
     public string? name { get; set; }
 
-    public decimal? price { get; set; }
+    public float? price { get; set; }
 
     public int unit_id { get; set; }
